requirements.txt
fastapi==0.111.0
pydantic==2.9.2
python-dotenv==1.0.1
transformers==4.45.2
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-gcp-trace